+++ src/app/(actions)/builds/update-linked-build.ts
@@ -4,6 +4,7 @@ import { LinkedBuild } from '@prisma/client'
 import { revalidatePath } from 'next/cache'
 
 import { getServerSession } from '@/app/(utils)/auth'
+import { cleanBadWords } from '@/app/(utils)/bad-word-filter'
 import { prisma } from '@/app/(utils)/db'
 import { validateLinkedBuild } from '@/app/(validators)/validate-linked-build'
 
@@ -11,6 +12,7 @@ type Props = {
   id: string
   createdById: string
   name: string
+  description: string
   linkedBuildItems: Array<{
     label: string
     buildId: string
@@ -55,7 +57,14 @@ export default async function updatedLinkedBuild(linkedBuild: Props): Promise<{
         id: linkedBuild.id,
       },
       data: {
-        name: linkedBuild.name,
+        name:
+          linkedBuild.name && linkedBuild.name !== ''
+            ? cleanBadWords(linkedBuild.name)
+            : '',
+        description:
+          linkedBuild.description && linkedBuild.description !== ''
+            ? cleanBadWords(linkedBuild.description)
+            : '',
         LinkedBuildItems: {
           create: linkedBuild.linkedBuildItems.map((linkedBuildItem) => ({
             createdAt: new Date(),