+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - When viewing a build, you can now click anywhere on the item button to view the item info. Previously this would only happen if you clicked the info icon.
+- Added description field to Linked Builds.
 
 ### Fixed